Test Cases Implemented

+++ WalletService/src/main/java/com/wallet/api/controller/WalletController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.wallet.api.model.Wallet;
@@ -19,18 +20,25 @@ public class WalletController {
 
 	@Autowired
 	WalletRepository walletRepository;
-	
-	@GetMapping("/wallet")
-	public List<Wallet> getString() {
+
+
+	@GetMapping("/walletS")// get the list of all wallets
+	public List<Wallet> getWallet() {
 		return walletRepository.findAll();
 	}
 	
-	@PostMapping("/wallet")
+	@GetMapping("/a")
+	public List<Wallet> a() {
+		return walletRepository.findAll();
+	}
+	
+	
+	@PostMapping("/wallet")// post the wallet
 	public Wallet postWallet(@RequestBody Wallet wallet) {
 		return walletRepository.save(wallet);
 	}
 	
-	@PutMapping("/wallet/{wId}")
+	@PutMapping("/wallet/{wId}")// edit the existent wallet
 	public Wallet putWallet(@PathVariable("wId") Long wId, @RequestBody Wallet wallet) {
 		Wallet wallet1=walletRepository.getById(wId);
 		
@@ -39,9 +47,10 @@ public class WalletController {
 		return walletRepository.save(wallet1);
 	}
 	
-	@DeleteMapping("/wallet/{wId}")
+	@DeleteMapping("/wallet/{wId}")// delete the existent wallet
 	public List<Wallet> deleteWallet(@PathVariable("wId") Long wId){
-		walletRepository.deleteById(wId);
+		Wallet wallet1=walletRepository.getById(wId);
+		if(wallet1.getId() !=0 ) walletRepository.deleteById(wId);
 		return walletRepository.findAll();
 	}
 }
